requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
gunicorn==21.2.0
werkzeug==2.3.7
marshmallow==3.20.1
